Review: one file per system, and the solver in its own
NetworkSystem.hpp was 2527 lines holding both systems and the Newton around
them. Split three ways, no code changed:

  NetworkSolve.hpp             Node, Result, DenseMatrix, shareByGuide, the
                               globalisations, Parameters and solve() -- what
                               neither system owns. 380 lines.
  NetworkInjectionSystem.hpp   Well, Control, InjectionSystem, dump/replay.
  NetworkProductionSystem.hpp  ProductionSystem, dump/replay.

solve() is a template over the system, so the solver header knows nothing about
either; both system headers include it. A stale comment above systemUsesAnalytic
that had been left behind by an earlier edit is dropped.

Tests and all three decks unchanged.
