+++ opm/simulators/wells/BlackoilWellModelNetworkGeneric.hpp
@@ -31,7 +31,8 @@
 
 #include <opm/simulators/flow/NewtonIterationContext.hpp>
 #include <opm/simulators/wells/NetworkNodePressureUpdater.hpp>
-#include <opm/simulators/wells/NetworkSystem.hpp>
+#include <opm/simulators/wells/NetworkInjectionSystem.hpp>
+#include <opm/simulators/wells/NetworkProductionSystem.hpp>
 #include <opm/simulators/utils/ParallelCommunication.hpp>
 
 #include <array>
