+++ opm/simulators/wells/NetworkProductionSystem.hpp
@@ -16,20 +16,16 @@
   You should have received a copy of the GNU General Public License
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
 */
-#ifndef OPM_NETWORK_SYSTEM_HEADER_INCLUDED
-#define OPM_NETWORK_SYSTEM_HEADER_INCLUDED
+#ifndef OPM_NETWORK_PRODUCTION_SYSTEM_HEADER_INCLUDED
+#define OPM_NETWORK_PRODUCTION_SYSTEM_HEADER_INCLUDED
 
-#include <opm/input/eclipse/EclipseState/Phase.hpp>
+#include <opm/simulators/wells/NetworkSolve.hpp>
+
 #include <opm/input/eclipse/Units/Units.hpp>
 
 #include <opm/simulators/wells/VFPHelpers.hpp>
-#include <opm/simulators/wells/VFPInjProperties.hpp>
 #include <opm/simulators/wells/VFPProdProperties.hpp>
 #include <opm/material/densead/Evaluation.hpp>
-
-#include <dune/common/dynmatrix.hh>
-#include <dune/common/dynvector.hh>
-#include <dune/common/fmatrix.hh>
 
 #include <algorithm>
 #include <array>
@@ -44,939 +40,6 @@
 #include <vector>
 
 namespace Opm::NetworkSolve {
-
-/// An injection network solved simultaneously in its pressures and its rates.
-///
-/// The unknowns are the pressure of every non-terminal node, the rate through
-/// every node's parent branch, each well's (rate, bhp), and a group multiplier
-/// when a target is active. The equations are the branch pressure drops, the
-/// node mass balances, each well's inflow performance and one control equation
-/// per well, plus the group target.
-///
-/// The alternative is to eliminate the rates and iterate on the node pressures
-/// alone, which is what the fixed-point and bracketing methods do. That residual
-/// is only piecewise differentiable -- the control limits put kinks in it -- and
-/// needs globalising. This one holds its controls fixed while a step is taken,
-/// so it is smooth within an active set and a plain Newton suffices.
-///
-/// Both the simulator and tests/test_networksolve.cpp fill this in; the wells
-/// differ (the simulator's inflow performance comes from the well Jacobian, the
-/// bench's from a reference operating point) but the system does not.
-
-constexpr int NoTable = 9999;   // GNETINJE's "no table": pressure passes through
-
-struct Node
-{
-    std::string name;
-    int parent = -1;            // -1 only for the terminal
-    int vfp_table = NoTable;
-    /// NEFAC: what this node passes on of what it collects.
-    double efficiency = 1.0;
-};
-
-template<class Scalar>
-struct Well
-{
-    std::string name;
-    int node = 0;
-    int vfp_table = 0;
-    /// Inflow performance in the simulator's convention, q = ipr_b * bhp - ipr_a
-    /// (WellState's implicit_ipr_a / implicit_ipr_b).
-    Scalar ipr_a = 0.0;
-    Scalar ipr_b = 0.0;
-    Scalar bhp_limit = 0.0;
-    Scalar rate_limit = 0.0;
-    /// Whether the group allocated this well. It counts against the group's
-    /// target whatever control it ends up on -- a well that runs into its own
-    /// bhp or rate limit still injects, and the wells that scale with the
-    /// multiplier have to make up the remainder, not the whole target.
-    bool in_group = false;
-    Scalar guide = 0.0;         // share of a group target
-    /// WEFAC as it applies to the network: the well's own rate is q, the branch
-    /// above it sees efficiency * q.
-    Scalar efficiency = 1.0;
-    /// Hydrostatic correction between the tubing table's datum and the well's
-    /// reference depth: the well's bhp is the table's less this.
-    Scalar vfp_dp = 0.0;
-    /// Rate to start the solve from. Zero means work one out from the tables,
-    /// which is all the bench can do; the simulator knows what the well is
-    /// actually doing and should say so, or the first control selection is made
-    /// on a rate that has nothing to do with the current state.
-    Scalar q_start = 0.0;
-};
-
-/// Which equation closes a well.
-enum class Control { Thp, Bhp, Rate, Grup };
-
-template<class Scalar>
-struct Result
-{
-    bool converged = false;
-    int iterations = 0;
-    std::vector<Scalar> node_pressure;   // every node, terminal included
-    std::vector<Scalar> well_rate;       // per well, in the order they were added
-
-    /// Why it stopped, for the caller to report. A converged solve leaves these
-    /// at the values that satisfied the test.
-    Scalar residual = 0.0;          // max norm of the scaled residual
-    bool controls_moving = false;   // an active-set change on the last iteration
-    bool guides_moving = false;     // group shares still settling
-    /// One letter per well per iteration for the last few iterations, so a
-    /// cycling active set can be read off: T thp, B bhp, R rate, G group.
-    std::string control_trace;
-    /// Iterations on which some well changed control. A solve that has to move
-    /// the active set a few times is working; one that keeps moving it is not.
-    int switches = 0;
-    /// Production only: every well's water/oil/gas and bhp at the solution.
-    std::vector<std::array<Scalar, 3>> well_phase_rates;
-    std::vector<Scalar> well_bhp;
-};
-
-/// Dense square system, solved by Dune. The networks this solves have tens of
-/// unknowns, so a dense direct solve is the whole story. Wrapped only to keep
-/// (i,j) indexing and to answer "singular" with false instead of an exception --
-/// the caller hands the network back to the relaxed update rather than throwing
-/// out of the well model.
-template<class Scalar>
-class DenseMatrix
-{
-public:
-    explicit DenseMatrix(const int n) : a_(n, n, Scalar{0}) {}
-
-    Scalar& operator()(const int i, const int j) { return a_[i][j]; }
-    Scalar operator()(const int i, const int j) const { return a_[i][j]; }
-
-    /// Solves A y = b. False if A is singular to working precision.
-    bool solve(const std::vector<Scalar>& b, std::vector<Scalar>& y) const
-    {
-        const auto n = a_.N();
-        Dune::DynamicVector<Scalar> rhs(n), x(n, Scalar{0});
-        std::copy(b.begin(), b.end(), rhs.begin());
-        try {
-            a_.solve(x, rhs);
-        } catch (const Dune::FMatrixError&) {
-            return false;
-        }
-        y.assign(x.begin(), x.end());
-        return true;
-    }
-
-private:
-    Dune::DynamicMatrix<Scalar> a_;
-};
-/// Divide a group target by guide rate, take out the wells whose own limits keep
-/// them below their share, and re-divide the rest among those that can take it.
-/// A well that is out gets no share at all, so its own control binds.
-///
-/// This is the fixed point an active set would otherwise have to find by
-/// iterating, and finding it here is what stops it cycling: a multiplier read
-/// from the iterate means "the even split" while nobody is on group control and
-/// "the remainder after the others' rates" while somebody is, and each of those
-/// two numbers selects the state that produces the other.
-template<class Scalar>
-std::vector<Scalar> shareByGuide(const std::vector<Scalar>& guide,
-                                 const std::vector<char>& in_group,
-                                 const std::vector<Scalar>& own,
-                                 const Scalar target)
-{
-    const int n = static_cast<int>(guide.size());
-    std::vector<Scalar> share(n, std::numeric_limits<Scalar>::max());
-    if (!(target > Scalar{0})) {
-        return share;
-    }
-
-    std::vector<char> pooled(in_group);
-    Scalar guides = 0.0;
-    for (int w = 0; w < n; ++w) {
-        if (pooled[w] != 0) {
-            guides += guide[w];
-        }
-    }
-
-    Scalar remaining = target;
-    for (int pass = 0; pass <= n; ++pass) {
-        if (!(guides > Scalar{0})) {
-            break;
-        }
-        int drop = -1;
-        Scalar worst = 0.0;
-        for (int w = 0; w < n; ++w) {
-            if (pooled[w] == 0) {
-                continue;
-            }
-            share[w] = guide[w] / guides * std::max(remaining, Scalar{0});
-            if (share[w] - own[w] > worst) {
-                worst = share[w] - own[w];
-                drop = w;
-            }
-        }
-        if (drop < 0) {
-            break;
-        }
-        pooled[drop] = 0;
-        guides -= guide[drop];
-        remaining -= own[drop];
-        share[drop] = std::numeric_limits<Scalar>::max();
-    }
-    return share;
-}
-
-
-/// An injection network solved as one system: the pressure of every non-terminal
-/// node, the rate through every node's parent branch, and each well's rate and
-/// bhp, with a group multiplier when a target is active. The equations are the
-/// branch pressure drop from the VFPINJ table, the node mass balance, the well's
-/// inflow performance, whichever of thp/bhp/rate/group closes the well, and the
-/// group total. The counterpart for a production network is ProductionSystem;
-/// the two differ in what a rate is -- one number here, three phases there.
-template<class Scalar>
-class InjectionSystem
-{
-public:
-    using State = std::vector<Scalar>;
-    using ScalarType = Scalar;
-
-    InjectionSystem(const VFPInjProperties<Scalar>& props, const Phase phase)
-        : props_(&props), phase_(phase)
-    {}
-
-    void addNode(Node n) { nodes_.push_back(std::move(n)); }
-    void addWell(Well<Scalar> w) { wells_.push_back(std::move(w)); }
-    void setTerminalPressure(const Scalar p) { terminal_pressure_ = p; }
-    /// The target of the one group this system can carry.
-    ///
-    /// A well belongs to that group by being under group control, not by where
-    /// it sits in the network: the group tree and the network tree are
-    /// independent and share only their leaves, and nothing here assumes
-    /// otherwise.
-    ///
-    /// What it does assume is a **single** constraining group. Two groups
-    /// binding different subsets of these wells would be summed into one target,
-    /// which is wrong, and nested groups need a multiplier each with a well's
-    /// share the product down its chain. Neither is modelled.
-    void setGroupTarget(const Scalar target) { group_target_ = target; }
-
-    /// Residual scale for the rate rows. Without one, rate and pressure rows
-    /// differ by several decades and no single tolerance means anything. The
-    /// default from finish() is a hundredth of the largest rate in play, which
-    /// is why it has to be called after the wells are in.
-    void setRateScale(const Scalar s) { rate_scale_ = s; }
-
-    /// Resolve the tree and the defaults. Call once everything is added.
-    void finish()
-    {
-        children_.assign(nodes_.size(), {});
-        wells_at_.assign(nodes_.size(), {});
-        for (std::size_t n = 1; n < nodes_.size(); ++n) {
-            children_[nodes_[n].parent].push_back(static_cast<int>(n));
-        }
-        for (std::size_t w = 0; w < wells_.size(); ++w) {
-            wells_at_[wells_[w].node].push_back(static_cast<int>(w));
-        }
-        for (auto& w : wells_) {
-            // A guide of nothing is a real answer for a well the group has put
-            // at zero -- it takes no share. Only fill one in when there is a
-            // rate limit to derive it from.
-            if (w.guide <= 0.0 && w.rate_limit > 0.0) {
-                w.guide = w.rate_limit;
-            }
-        }
-        if (rate_scale_ <= 0.0) {
-            Scalar largest = group_target_;
-            for (const auto& w : wells_) {
-                largest = std::max(largest, w.rate_limit);
-            }
-            rate_scale_ = std::max(largest * Scalar{0.01},
-                                   unit::convert::from(1.0, unit::cubic(unit::meter) / unit::day));
-        }
-        controls_.assign(wells_.size(), Control::Thp);
-        for (std::size_t w = 0; w < wells_.size(); ++w) {
-            if (grouped() && wells_[w].in_group) {
-                controls_[w] = Control::Grup;
-            }
-        }
-    }
-
-    int numNodes() const { return static_cast<int>(nodes_.size()) - 1; }
-    int numWells() const { return static_cast<int>(wells_.size()); }
-    bool grouped() const { return group_target_ > 0.0; }
-
-    std::vector<Scalar> guides() const
-    {
-        std::vector<Scalar> g(wells_.size());
-        std::transform(wells_.begin(), wells_.end(), g.begin(),
-                       [](const auto& w) { return w.guide; });
-        return g;
-    }
-
-    std::vector<char> inGroup() const
-    {
-        std::vector<char> in(wells_.size());
-        std::transform(wells_.begin(), wells_.end(), in.begin(),
-                       [](const auto& w) { return static_cast<char>(w.in_group); });
-        return in;
-    }
-    int size() const { return 2 * numNodes() + 2 * numWells() + (grouped() ? 1 : 0); }
-
-    Phase phase() const { return phase_; }
-    Scalar terminalPressure() const { return terminal_pressure_; }
-    Scalar groupTarget() const { return group_target_; }
-    const std::vector<Node>& nodes() const { return nodes_; }
-    const std::vector<Well<Scalar>>& wells() const { return wells_; }
-    Control control(const int w) const { return controls_[w]; }
-
-    /// One letter for the trace a failed solve reports.
-    char controlLetter(const int w) const
-    {
-        switch (controls_[w]) {
-        case Control::Thp:  return 'T';
-        case Control::Bhp:  return 'B';
-        case Control::Rate: return 'R';
-        case Control::Grup: return 'G';
-        }
-        return '?';
-    }
-
-    int pIdx(const int node) const { return node - 1; }
-    int qIdx(const int node) const { return numNodes() + node - 1; }
-    int qwIdx(const int w) const { return 2 * numNodes() + w; }
-    int bhpIdx(const int w) const { return 2 * numNodes() + numWells() + w; }
-    int lambdaIdx() const { return 2 * numNodes() + 2 * numWells(); }
-
-    bool hasTable(const Node& n) const { return n.vfp_table != NoTable; }
-
-    /// Below this a table lookup has not answered, it has run out of table.
-    static constexpr Scalar kTableFloor = unit::barsa;
-
-    static Scalar ipr(const Well<Scalar>& w, const Scalar bhp) { return w.ipr_b * bhp - w.ipr_a; }
-
-    /// Clamp table lookups to the axes, as the fixed-point pressure computation
-    /// does. Leave this off for a Newton: outside the box the residual then goes
-    /// flat and the Jacobian is singular in the rates, so there is nothing to
-    /// descend. limitStep() is the treatment that works. It exists here only so
-    /// that the comparison can be made -- see test_networksolve.cpp.
-    void setClampToAxes(const bool on) { clamp_to_axes_ = on; }
-
-    /// A table lookup with the two derivatives the Jacobian needs. They come
-    /// free with the interpolation and are otherwise thrown away.
-    struct Lookup
-    {
-        Scalar value = 0.0;
-        Scalar dthp = 0.0;    // d(bhp)/d(thp)
-        Scalar dflo = 0.0;    // d(bhp)/d(rate)
-    };
-
-    Lookup tableLookup(const int table, const Scalar thp, const Scalar q_in) const
-    {
-        Scalar q = q_in;
-        Scalar p = thp;
-        const auto& t = props_->getTable(table);
-        bool clamped_flo = false;
-        bool clamped_thp = false;
-        if (clamp_to_axes_) {
-            const Scalar lo = t.getFloAxis().front(), hi = t.getFloAxis().back();
-            const Scalar plo = t.getTHPAxis().front(), phi = t.getTHPAxis().back();
-            clamped_flo = (q < lo) || (q > hi);
-            clamped_thp = (p < plo) || (p > phi);
-            q = std::clamp(q, lo, hi);
-            p = std::clamp(p, plo, phi);
-        }
-        const Scalar aqua = (phase_ == Phase::WATER) ? q : Scalar{0};
-        const Scalar vapour = (phase_ == Phase::GAS) ? q : Scalar{0};
-        const auto e = VFPHelpers<Scalar>::bhp(t, aqua, liquid_, vapour, p);
-        // Where the lookup was clamped the value no longer moves with the input,
-        // which is exactly the flat residual that makes clamping a bad idea for
-        // a Newton -- but the derivative has to report it honestly.
-        return {e.value, clamped_thp ? Scalar{0} : e.dthp, clamped_flo ? Scalar{0} : e.dflo};
-    }
-
-    /// Downstream pressure of a branch, or a well's bhp: the same table lookup.
-    Scalar tableBhp(const int table, const Scalar thp, const Scalar q_in) const
-    {
-        Scalar q = q_in;
-        Scalar p = thp;
-        if (clamp_to_axes_) {
-            const auto& t = props_->getTable(table);
-            q = std::clamp(q, t.getFloAxis().front(), t.getFloAxis().back());
-            p = std::clamp(p, t.getTHPAxis().front(), t.getTHPAxis().back());
-        }
-        const Scalar aqua = (phase_ == Phase::WATER) ? q : Scalar{0};
-        const Scalar vapour = (phase_ == Phase::GAS) ? q : Scalar{0};
-        return props_->bhp(table, aqua, Scalar{0}, vapour, p);
-    }
-
-    /// Rate this well would take on THP control at a given node pressure: its
-    /// inflow performance met with its tubing curve, then its own limits.
-    ///
-    /// This is the well's capability at a network pressure, which is what a
-    /// share of a group target should be proportional to. Its current rate is
-    /// not: that is the split one is trying to decide, so using it as the guide
-    /// makes the allocation reproduce whatever it already was.
-    /// The rate thp control allows at this node pressure.
-    ///
-    /// `cap_by_rate_limit` is the whole subtlety. Bounding the search by the
-    /// well's own rate limit makes thp's allowance tie with that limit and win
-    /// the tie, so the well stays on thp -- whose equation says nothing about a
-    /// rate -- and can settle above its limit. Removing the bound fixes that and
-    /// costs far more than it buys: while the pressures are still moving, a
-    /// well's crossing routinely lies past its limit, rate control pins it
-    /// there, four wells pinned at their limits ask the network for several
-    /// times what it carries, and the globalisation basin falls from 511/529 to
-    /// 271/529. So the bound stays on while the solve is still moving, and
-    /// solve() drops it once there is a converged point to enforce the limit
-    /// from -- an iterate that is no longer transient.
-    Scalar thpPotential(const Well<Scalar>& w, const Scalar p_node,
-                        const bool cap_by_rate_limit = true) const
-    {
-        const auto& t = props_->getTable(w.vfp_table);
-        const auto& axis = t.getFloAxis();
-        const Scalar lo = axis.front();
-        Scalar hi = (cap_by_rate_limit && w.rate_limit > Scalar{0})
-            ? std::min(w.rate_limit, axis.back()) : axis.back();
-        if (!cap_by_rate_limit) {
-            // These tables are padded with zeros past the rates they describe,
-            // and a bhp of nothing is not a bhp. Walk back to the last rate this
-            // one answers for; a root past that is a root in the padding.
-            for (std::size_t i = axis.size(); i-- > 0;) {
-                if (axis[i] > lo && tableBhp(w.vfp_table, p_node, axis[i]) > kTableFloor) {
-                    hi = axis[i];
-                    break;
-                }
-            }
-        }
-        if (!(hi > lo)) {
-            return Scalar{0};
-        }
-        // bhp falls with rate at fixed thp in these tables, so f is decreasing.
-        const auto f = [&](const Scalar q) {
-            return ipr(w, tableBhp(w.vfp_table, p_node, q) - w.vfp_dp) - q;
-        };
-        if (f(lo) <= Scalar{0}) {
-            return Scalar{0};
-        }
-        if (f(hi) >= Scalar{0}) {
-            return hi;
-        }
-        Scalar a = lo, b = hi, q = hi;
-        for (int it = 0; it < 60; ++it) {
-            q = Scalar{0.5} * (a + b);
-            (f(q) > Scalar{0} ? a : b) = q;
-        }
-        return std::max(q, Scalar{0});
-    }
-
-    /// Stop capping thp's allowance with each well's rate limit, so a well whose
-    /// tubing would carry more than it is allowed goes on rate control. Only
-    /// safe from a converged iterate -- see thpPotential().
-    void setEnforceRateLimits(const bool on) { enforce_rate_limits_ = on; }
-
-    /// Any well come to rest above its own rate limit.
-    bool rateLimitsViolated(const State& x) const
-    {
-        for (int w = 0; w < numWells(); ++w) {
-            if (wells_[w].rate_limit > Scalar{0}
-                && x[qwIdx(w)] > wells_[w].rate_limit * (Scalar{1} + Scalar{1e-9})) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /// Take the guide rates from thpPotential() at the current node pressures
-    /// instead of whatever the caller supplied. Only meaningful with a group
-    /// target, and only when the caller has no better guide of its own.
-    void setGuidesFromPotential(const bool on) { guides_from_potential_ = on; }
-    bool guidesFromPotential() const { return guides_from_potential_; }
-
-    /// Recompute the guides from the current iterate. Returns the largest
-    /// relative change, so the caller can tell when they have settled.
-    Scalar refreshGuides(const State& x)
-    {
-        if (!guides_from_potential_ || !grouped()) {
-            return Scalar{0};
-        }
-        Scalar moved = 0.0;
-        for (auto& w : wells_) {
-            const Scalar p = (w.node == 0) ? terminal_pressure_ : x[pIdx(w.node)];
-            const Scalar potential = thpPotential(w, p);
-            if (potential > Scalar{0}) {
-                moved = std::max(moved, std::abs(potential - w.guide) / std::max(w.guide, potential));
-                w.guide = potential;
-            }
-        }
-        return moved;
-    }
-
-    /// Largest rate the table describes. Past it the cells are zero-filled and
-    /// the interpolation runs away, so this is the edge of the feasible set.
-    Scalar maxFlow(const int table) const { return props_->getTable(table).getFloAxis().back(); }
-
-    State residual(const State& x) const
-    {
-        const int nodes = numNodes();
-        const int wells = numWells();
-        State r(size(), 0.0);
-
-        auto pressure = [&](const int n) { return n == 0 ? terminal_pressure_ : x[pIdx(n)]; };
-
-        for (int n = 1; n <= nodes; ++n) {
-            const auto& node = nodes_[n];
-            const Scalar upstream = pressure(node.parent);
-            r[n - 1] = hasTable(node)
-                ? x[pIdx(n)] - tableBhp(node.vfp_table, upstream, x[qIdx(n)])
-                : x[pIdx(n)] - upstream;
-
-            Scalar balance = x[qIdx(n)];
-            for (const int c : children_[n]) {
-                balance -= nodes_[c].efficiency * x[qIdx(c)];
-            }
-            for (const int w : wells_at_[n]) {
-                balance -= wells_[w].efficiency * x[qwIdx(w)];
-            }
-            r[nodes + n - 1] = balance;
-        }
-
-        Scalar injected = 0.0;
-        for (int w = 0; w < wells; ++w) {
-            const auto& well = wells_[w];
-            const Scalar q = x[qwIdx(w)];
-            const Scalar bhp = x[bhpIdx(w)];
-            // Every well the group allocated counts against the target, on
-            // whatever control it ended up. Counting only those still on group
-            // control asks the rest to deliver the whole target while a limited
-            // well injects on top of it, and the group over-delivers by exactly
-            // that well's rate. Counting wells the group never allocated is the
-            // opposite error and cannot be satisfied at all.
-            if (well.in_group) {
-                injected += q;
-            }
-
-            r[2 * nodes + w] = (q - ipr(well, bhp)) / rate_scale_;
-
-            Scalar& control = r[2 * nodes + wells + w];
-            switch (controls_[w]) {
-            case Control::Thp:
-                control = (bhp - (tableBhp(well.vfp_table, pressure(well.node), q) - well.vfp_dp))
-                        / pressure_scale_;
-                break;
-            case Control::Bhp:
-                control = (bhp - well.bhp_limit) / pressure_scale_;
-                break;
-            case Control::Rate:
-                control = (q - well.rate_limit) / rate_scale_;
-                break;
-            case Control::Grup:
-                control = (q - well.guide * x[lambdaIdx()]) / rate_scale_;
-                break;
-            }
-        }
-
-        if (grouped()) {
-            // With nobody on group control the multiplier is free, so pin it
-            // rather than hand the Newton a singular column.
-            const bool any = std::find(controls_.begin(), controls_.end(), Control::Grup)
-                           != controls_.end();
-            r[lambdaIdx()] = any ? (injected - group_target_) / rate_scale_
-                                 : (x[lambdaIdx()] - lambda0()) / rate_scale_;
-        }
-
-        for (int n = 0; n < nodes; ++n) {
-            r[n] /= pressure_scale_;
-            r[nodes + n] /= rate_scale_;
-        }
-        return r;
-    }
-
-    /// Take the last well out of the group, so a test can build a network whose
-    /// group does not hold every well on it.
-    void dropLastFromGroup()
-    {
-        if (!wells_.empty()) {
-            wells_.back().in_group = false;
-        }
-    }
-
-    /// Reselect each well's control: the most restrictive violated limit wins,
-    /// the same rule a clamp would apply. Choosing by a fixed priority instead
-    /// makes the active set chatter and the Newton never terminates. Returns
-    /// true if anything moved -- converging with a control still moving is not
-    /// converged.
-    bool updateControls(const State& x)
-    {
-        const int n = numWells();
-
-        // What each well could inject on a control of its own, at this iterate's
-        // node pressures. Nothing here reads the iterate's q, bhp or multiplier:
-        // mid-Newton those are not a consistent well state, on rate control q
-        // *is* the limit, and the multiplier is defined by the very active set
-        // being chosen here.
-        std::vector<Scalar> own(n), thp(n);
-        for (int w = 0; w < n; ++w) {
-            const auto& well = wells_[w];
-            const Scalar p_node = (well.node == 0) ? terminal_pressure_ : x[pIdx(well.node)];
-            thp[w] = thpPotential(well, p_node, !enforce_rate_limits_);
-            own[w] = std::min(thp[w], ipr(well, well.bhp_limit));
-            if (well.rate_limit > Scalar{0}) {
-                own[w] = std::min(own[w], well.rate_limit);
-            }
-        }
-
-        const auto share = shareByGuide(guides(), inGroup(), own, group_target_);
-
-        bool changed = false;
-        for (int w = 0; w < n; ++w) {
-            const auto& well = wells_[w];
-
-            // Given the node pressure, every control determines the well
-            // completely and so names the rate it would allow. The binding one
-            // is simply the smallest: nothing is "violated", and a control stops
-            // binding by being overtaken, which is how a well leaves one.
-            auto wanted = Control::Thp;
-            Scalar smallest = thp[w];
-            auto consider = [&](const Control c, const Scalar allows) {
-                if (allows < smallest) {
-                    smallest = allows;
-                    wanted = c;
-                }
-            };
-            consider(Control::Bhp, ipr(well, well.bhp_limit));
-            // A well the deck gives no rate limit is not a well limited to
-            // nothing; rate control simply has nothing to say about it.
-            if (well.rate_limit > Scalar{0}) {
-                consider(Control::Rate, well.rate_limit);
-            }
-            if (grouped() && well.in_group) {
-                consider(Control::Grup, share_from_multiplier_ ? well.guide * x[lambdaIdx()]
-                                                               : share[w]);
-            }
-
-            changed |= (wanted != controls_[w]);
-            controls_[w] = wanted;
-        }
-        return changed;
-    }
-
-    /// A starting point derived from a guess at every node's pressure.
-    State start(const State& node_pressure) const
-    {
-        State x(size(), 0.0);
-        std::vector<Scalar> well_rate(numWells());
-        for (int n = 1; n <= numNodes(); ++n) {
-            x[pIdx(n)] = node_pressure[n];
-        }
-        for (int w = 0; w < numWells(); ++w) {
-            const auto& well = wells_[w];
-            const Scalar guess = well.q_start > Scalar{0}
-                ? well.q_start : std::max(well.rate_limit * Scalar{0.1}, rate_scale_);
-            x[bhpIdx(w)] = tableBhp(well.vfp_table, node_pressure[well.node], guess);
-            // Deliberately just inside the limit, never exactly on it: the
-            // control tests below are inclusive, so opening on the limit would
-            // put every well on rate control before the solve has begun.
-            const Scalar most = Scalar{0.999} * well.rate_limit;
-            x[qwIdx(w)] = well.q_start > Scalar{0}
-                ? std::min(well.q_start, most)
-                : std::clamp(ipr(well, x[bhpIdx(w)]), Scalar{0}, most);
-            well_rate[w] = x[qwIdx(w)];
-        }
-        for (int n = numNodes(); n >= 1; --n) {
-            Scalar q = 0.0;
-            for (const int w : wells_at_[n]) {
-                q += wells_[w].efficiency * well_rate[w];
-            }
-            for (const int c : children_[n]) {
-                q += nodes_[c].efficiency * x[qIdx(c)];
-            }
-            x[qIdx(n)] = q;
-        }
-        if (grouped()) {
-            x[lambdaIdx()] = lambda0();
-        }
-        return x;
-    }
-
-    /// Rate of every well, in the order they were added.
-    State wellRates(const State& x) const
-    {
-        State q(wells_.size());
-        for (int w = 0; w < numWells(); ++w) {
-            q[w] = x[qwIdx(w)];
-        }
-        return q;
-    }
-
-    /// Pressure at every node, terminal included.
-    State pressures(const State& x) const
-    {
-        State p(nodes_.size(), terminal_pressure_);
-        for (int n = 1; n <= numNodes(); ++n) {
-            p[n] = x[pIdx(n)];
-        }
-        return p;
-    }
-
-    /// Natural magnitude of unknown i, so one step cap or trust radius can apply
-    /// to a vector holding both pressures and rates.
-    Scalar columnScale(const int i) const
-    {
-        const bool is_pressure = (i < numNodes()) || (i >= bhpIdx(0) && i < lambdaIdx());
-        return is_pressure ? pressure_scale_ : rate_scale_;
-    }
-
-    /// Keep the branch flows inside the box the tables describe, by projecting
-    /// the offending components rather than scaling all of the step -- one
-    /// binding rate should not throttle the pressure updates too. Only the
-    /// branch flows: a well's own rate limit already has a control equation, and
-    /// bounding it would stop that control ever activating.
-    State limitStep(const State& x, const State& dx) const
-    {
-        State limited = dx;
-        for (int n = 1; n <= numNodes(); ++n) {
-            const auto& node = nodes_[n];
-            if (!hasTable(node)) {
-                continue;
-            }
-            const int i = qIdx(n);
-            const Scalar hi = maxFlow(node.vfp_table);
-            if (x[i] <= hi && x[i] + limited[i] > hi) {
-                limited[i] = hi - x[i];
-            }
-            if (x[i] >= 0.0 && x[i] + limited[i] < 0.0) {
-                limited[i] = -x[i];
-            }
-        }
-        return limited;
-    }
-
-    Scalar pressureScale() const { return pressure_scale_; }
-
-    /// Assemble the Jacobian from the table derivatives instead of differencing
-    /// the residual. Everything but the two branch/tubing lookups is constant,
-    /// so this is n+1 residual evaluations replaced by one pass.
-    void setAnalyticJacobian(const bool on) { analytic_jacobian_ = on; }
-    bool usesAnalyticJacobian() const { return analytic_jacobian_; }
-
-
-    /// Take a well's group share from the iterate's multiplier instead of
-    /// resolving the split. This is the rule that cycles, kept so the bench can
-    /// measure the two on the same systems; nothing should turn it on.
-    void setGroupShareFromMultiplier(const bool on) { share_from_multiplier_ = on; }
-
-    /// The Jacobian of residual() at x, entry by entry.
-    DenseMatrix<Scalar> jacobian(const State& x) const
-    {
-        const int nodes = numNodes();
-        const int wells = numWells();
-        DenseMatrix<Scalar> J(size());
-
-        auto pressure = [&](const int n) { return n == 0 ? terminal_pressure_ : x[pIdx(n)]; };
-        // Row i is divided by scale(i) in residual(), so its derivatives are too.
-        auto add = [&](const int row, const int col, const Scalar value, const Scalar scale) {
-            J(row, col) += value / scale;
-        };
-
-        for (int n = 1; n <= nodes; ++n) {
-            const auto& node = nodes_[n];
-            const int row = n - 1;
-            add(row, pIdx(n), 1.0, pressure_scale_);
-            if (hasTable(node)) {
-                const auto e = tableLookup(node.vfp_table, pressure(node.parent), x[qIdx(n)]);
-                if (node.parent != 0) {
-                    add(row, pIdx(node.parent), -e.dthp, pressure_scale_);
-                }
-                add(row, qIdx(n), -e.dflo, pressure_scale_);
-            } else if (node.parent != 0) {
-                add(row, pIdx(node.parent), -1.0, pressure_scale_);
-            }
-
-            const int balance = nodes + n - 1;
-            add(balance, qIdx(n), 1.0, rate_scale_);
-            for (const int c : children_[n]) {
-                add(balance, qIdx(c), -nodes_[c].efficiency, rate_scale_);
-            }
-            for (const int w : wells_at_[n]) {
-                add(balance, qwIdx(w), -wells_[w].efficiency, rate_scale_);
-            }
-        }
-
-        for (int w = 0; w < wells; ++w) {
-            const auto& well = wells_[w];
-            const int ipr_row = 2 * nodes + w;
-            add(ipr_row, qwIdx(w), 1.0, rate_scale_);
-            add(ipr_row, bhpIdx(w), -well.ipr_b, rate_scale_);
-
-            const int row = 2 * nodes + wells + w;
-            switch (controls_[w]) {
-            case Control::Thp: {
-                const auto e = tableLookup(well.vfp_table, pressure(well.node), x[qwIdx(w)]);
-                add(row, bhpIdx(w), 1.0, pressure_scale_);
-                if (well.node != 0) {
-                    add(row, pIdx(well.node), -e.dthp, pressure_scale_);
-                }
-                add(row, qwIdx(w), -e.dflo, pressure_scale_);
-                break;
-            }
-            case Control::Bhp:
-                add(row, bhpIdx(w), 1.0, pressure_scale_);
-                break;
-            case Control::Rate:
-                add(row, qwIdx(w), 1.0, rate_scale_);
-                break;
-            case Control::Grup:
-                add(row, qwIdx(w), 1.0, rate_scale_);
-                add(row, lambdaIdx(), -well.guide, rate_scale_);
-                break;
-            }
-        }
-
-        if (grouped()) {
-            const bool any = std::find(controls_.begin(), controls_.end(), Control::Grup)
-                           != controls_.end();
-            if (any) {
-                // The group's own wells, matching the residual. Differentiating
-                // every well instead is a wrong derivative that no bench case
-                // could see, because there every well is in the group.
-                for (int w = 0; w < wells; ++w) {
-                    if (wells_[w].in_group) {
-                        add(lambdaIdx(), qwIdx(w), 1.0, rate_scale_);
-                    }
-                }
-            } else {
-                add(lambdaIdx(), lambdaIdx(), 1.0, rate_scale_);
-            }
-        }
-        return J;
-    }
-
-private:
-    Scalar lambda0() const
-    {
-        Scalar guides = 0.0;
-        for (const auto& w : wells_) {
-            guides += w.guide;
-        }
-        return guides > 0.0 ? group_target_ / guides : Scalar{0};
-    }
-
-    const VFPInjProperties<Scalar>* props_;
-    Phase phase_;
-    std::vector<Node> nodes_;
-    std::vector<Well<Scalar>> wells_;
-    std::vector<std::vector<int>> children_;
-    std::vector<std::vector<int>> wells_at_;
-    std::vector<Control> controls_;
-
-    Scalar terminal_pressure_ = 0.0;
-    Scalar group_target_ = 0.0;
-    Scalar rate_scale_ = 0.0;
-    Scalar liquid_ = 0.0;
-    bool clamp_to_axes_ = false;
-    bool analytic_jacobian_ = false;
-    bool share_from_multiplier_ = false;
-    bool enforce_rate_limits_ = false;
-    bool guides_from_potential_ = false;
-    Scalar pressure_scale_ = unit::barsa;
-};
-
-/// Write everything the solve works from, so a failure can be replayed offline.
-/// The VFP tables are not included -- the reader supplies them from the deck.
-template<class Scalar>
-void write(const InjectionSystem<Scalar>& system, const std::vector<Scalar>& guess, std::ostream& os)
-{
-    os << "phase " << (system.phase() == Phase::GAS ? "GAS" : "WATER") << '\n'
-       << "terminal " << system.terminalPressure() << '\n'
-       << "group_target " << system.groupTarget() << '\n'
-       << "guides_from_potential " << system.guidesFromPotential() << '\n'
-       << "analytic_jacobian " << system.usesAnalyticJacobian() << '\n';
-    for (const auto& n : system.nodes()) {
-        os << "node " << n.name << ' ' << n.parent << ' ' << n.vfp_table << ' '
-           << n.efficiency << '\n';
-    }
-    for (const auto& w : system.wells()) {
-        os << "well " << w.name << ' ' << w.node << ' ' << w.vfp_table << ' '
-           << w.ipr_a << ' ' << w.ipr_b << ' ' << w.bhp_limit << ' '
-           << w.rate_limit << ' ' << w.guide << ' ' << w.q_start << ' '
-           << w.in_group << ' ' << w.efficiency << ' ' << w.vfp_dp << '\n';
-    }
-    os << "guess";
-    for (const auto p : guess) {
-        os << ' ' << p;
-    }
-    os << '\n';
-}
-
-/// Rebuild a written system against tables the caller already has. Returns the
-/// system and the starting pressures it was given.
-template<class Scalar>
-std::pair<InjectionSystem<Scalar>, std::vector<Scalar>>
-read(std::istream& is, const VFPInjProperties<Scalar>& props)
-{
-    std::string tag;
-    Phase phase = Phase::GAS;
-    Scalar terminal = 0.0, target = 0.0;
-    bool guides_from_potential = false, analytic_jacobian = false;
-    std::vector<Node> nodes;
-    std::vector<Well<Scalar>> wells;
-    std::vector<Scalar> guess;
-
-    std::string line;
-    while (std::getline(is, line)) {
-        std::istringstream in(line);
-        if (!(in >> tag)) {
-            continue;
-        }
-        if (tag == "phase") {
-            std::string name;
-            in >> name;
-            phase = (name == "GAS") ? Phase::GAS : Phase::WATER;
-        } else if (tag == "terminal") {
-            in >> terminal;
-        } else if (tag == "group_target") {
-            in >> target;
-        } else if (tag == "node") {
-            Node n;
-            in >> n.name >> n.parent >> n.vfp_table;
-            in >> n.efficiency;          // older dumps: stays 1
-            nodes.push_back(std::move(n));
-        } else if (tag == "well") {
-            Well<Scalar> w;
-            in >> w.name >> w.node >> w.vfp_table >> w.ipr_a >> w.ipr_b
-               >> w.bhp_limit >> w.rate_limit >> w.guide >> w.q_start;
-            // Without this a replay solves an ungrouped system -- a different,
-            // easier problem than the one that failed. Older dumps have no
-            // field; take them as fully grouped, which is what they were.
-            int grouped = 1;
-            in >> grouped;
-            w.in_group = (grouped != 0);
-            in >> w.efficiency;          // older dumps: stays 1
-            in >> w.vfp_dp;              // older dumps: stays 0
-            wells.push_back(std::move(w));
-        } else if (tag == "guides_from_potential") {
-            in >> guides_from_potential;
-        } else if (tag == "analytic_jacobian") {
-            in >> analytic_jacobian;
-        } else if (tag == "guess") {
-            Scalar p;
-            while (in >> p) {
-                guess.push_back(p);
-            }
-        }
-    }
-
-    InjectionSystem<Scalar> system(props, phase);
-    system.setTerminalPressure(terminal);
-    system.setGroupTarget(target);
-    system.setGuidesFromPotential(guides_from_potential);
-    system.setAnalyticJacobian(analytic_jacobian);
-    for (auto& n : nodes) {
-        system.addNode(std::move(n));
-    }
-    for (auto& w : wells) {
-        system.addWell(std::move(w));
-    }
-    system.finish();
-    return {std::move(system), std::move(guess)};
-}
-
 
 // ---------------------------------------------------------------------------
 // Production networks
@@ -2300,218 +1363,6 @@ readProduction(std::istream& is, const VFPProdProperties<Scalar>& props, const U
     return {std::move(system), std::move(guess)};
 }
 
-/// Take the Newton step as it comes. This is what the full system wants: it has
-/// no kinks within an active set, so there is nothing for a globalisation to fix.
-struct FullStep
-{
-    template<class Sys, class State>
-    State accept(const Sys&, const State& x, const State&, const State& dx) const
-    {
-        State next(x.size());
-        for (std::size_t i = 0; i < x.size(); ++i) {
-            next[i] = x[i] + dx[i];
-        }
-        return next;
-    }
-};
-
-/// Backtrack until the residual norm drops. Not needed on the full system, but
-/// useful when the residual is not smooth.
-struct LineSearch
-{
-    int max_halvings = 12;
-
-    template<class Sys, class State>
-    State accept(const Sys& system, const State& x, const State& r, const State& dx) const
-    {
-        auto norm2 = [](const State& v) {
-            double s = 0.0;
-            for (const auto e : v) {
-                s += e * e;
-            }
-            return std::sqrt(s);
-        };
-        const double f0 = norm2(r);
-        double lambda = 1.0;
-        State trial(x.size());
-        for (int k = 0; k < max_halvings; ++k) {
-            for (std::size_t i = 0; i < x.size(); ++i) {
-                trial[i] = x[i] + lambda * dx[i];
-            }
-            if (norm2(system.residual(trial)) < f0) {
-                return trial;
-            }
-            lambda *= 0.5;
-        }
-        return trial;
-    }
-};
-
-/// Solve the system from a guess at the node pressures. The tolerance is on the
-/// scaled residual, so it reads as bar on the pressure rows.
-/// Ask a system whether it assembles its own Jacobian, without requiring that
-/// every system knows how.
-template<class Sys>
-bool systemUsesAnalytic(const Sys& system)
-{
-    if constexpr (requires { system.usesAnalyticJacobian(); }) {
-        return system.usesAnalyticJacobian();
-    } else {
-        return false;
-    }
-}
-
-template<class Sys, class State>
-auto systemJacobian(const Sys& system, const State& x)
-{
-    if constexpr (requires { system.jacobian(x); }) {
-        return system.jacobian(x);
-    } else {
-        return DenseMatrix<typename Sys::ScalarType>(system.size());
-    }
-}
-
-/// Convergence settings for solve(). No defaults: a caller states what it wants.
-template<class Scalar>
-struct Parameters
-{
-    /// Max norm of the scaled residual at which the system is converged.
-    Scalar tolerance;
-    int max_iterations;
-};
-
-/// Solve an InjectionSystem or a ProductionSystem by Newton-Raphson, choosing
-/// each well's control by an active set as it goes. Takes the node pressures to
-/// start from; returns the converged pressures and rates, or a Result with
-/// converged false and the reason it stopped.
-template<class Sys, class Globalisation>
-Result<typename Sys::ScalarType>
-solve(Sys& system,
-      const std::vector<typename Sys::ScalarType>& node_pressure_guess,
-      const Parameters<typename Sys::ScalarType> params,
-      Globalisation globalisation)
-{
-    const auto tolerance = params.tolerance;
-    const int max_iterations = params.max_iterations;
-    using Scalar = typename Sys::ScalarType;
-    auto x = system.start(node_pressure_guess);
-    const int n = system.size();
-    Result<Scalar> last;
-    std::vector<std::string> trace;
-    auto joined = [&trace] {
-        std::string out;
-        for (const auto& e : trace) {
-            out += (out.empty() ? "" : " ") + e;
-        }
-        return out;
-    };
-
-    // Only under --network-group-control, where the network places the group's
-    // split itself: the share is each well's potential at the node pressure, so
-    // it cannot be known before the starting pressures. Explicit, like OPM's own
-    // guide rates, and for the same reason -- recomputing it every iteration
-    // makes each share a moving target while its rate chases it, and the active
-    // set cycles between group and thp control. Making it implicit (the share an
-    // unknown of the system) or iterating it to a fixed point in an outer loop
-    // are the ways past that; both are open.
-    if constexpr (requires { system.refreshGuides(x); }) {
-        system.refreshGuides(x);
-    }
-
-    int switches = 0;
-    bool enforcing = false;
-    for (int it = 1; it <= max_iterations; ++it) {
-        const bool controls_moved = system.updateControls(x);
-        switches += controls_moved ? 1 : 0;
-        const auto r = system.residual(x);
-
-        Scalar worst = 0.0;
-        for (const auto e : r) {
-            worst = std::max(worst, std::abs(e));
-        }
-        {   // remember the active set, so a cycle can be seen in the report
-            std::string set;
-            for (int w = 0; w < system.numWells(); ++w) {
-                set += system.controlLetter(w);
-            }
-            trace.push_back(set);
-            if (trace.size() > 8) {
-                trace.erase(trace.begin());
-            }
-        }
-        const bool settled = !controls_moved;
-        if (worst < tolerance && settled) {
-            // Converged, but possibly with a well parked above its own rate
-            // limit -- thp's allowance is capped by that limit while the solve
-            // is moving, and a capped allowance ties with it. Now that the
-            // iterate is not transient, drop the cap and carry on from here;
-            // whoever is over the line goes on rate control and the rest take
-            // it up. The `enforcing` flag below is what keeps this to one pass,
-            // so a well that is still over the line afterwards converges as it
-            // is rather than dropping the cap again.
-            if constexpr (requires { system.setEnforceRateLimits(true); }) {
-                if (!enforcing && system.rateLimitsViolated(x)) {
-                    system.setEnforceRateLimits(true);
-                    enforcing = true;
-                    continue;
-                }
-            }
-            Result<Scalar> done{true, it, system.pressures(x), system.wellRates(x), worst,
-                                false, false, {}, switches};
-            if constexpr (requires { system.wellPhaseRates(x); system.wellBhps(x); }) {
-                done.well_phase_rates = system.wellPhaseRates(x);
-                done.well_bhp = system.wellBhps(x);
-            }
-            return done;
-        }
-        last = {false, it, {}, {}, worst, controls_moved, false, joined(), switches};
-
-        // A system that can hand over an assembled Jacobian does; the rest are
-        // differenced. The production prototype has no analytic one yet.
-        DenseMatrix<Scalar> J = systemUsesAnalytic(system)
-            ? systemJacobian(system, x)
-            : [&] {
-                  DenseMatrix<Scalar> fd(n);
-                  for (int j = 0; j < n; ++j) {
-                      auto shifted = x;
-                      const Scalar h = 1e-2 * system.columnScale(j);
-                      shifted[j] += h;
-                      const auto rj = system.residual(shifted);
-                      for (int i = 0; i < n; ++i) {
-                          fd(i, j) = (rj[i] - r[i]) / h;
-                      }
-                  }
-                  return fd;
-              }();
-
-        std::vector<Scalar> negative(n), dx;
-        for (int i = 0; i < n; ++i) {
-            negative[i] = -r[i];
-        }
-        if (!J.solve(negative, dx)) {
-            last.node_pressure = system.pressures(x);
-            last.well_rate = system.wellRates(x);
-            return last;
-        }
-
-        dx = system.limitStep(x, dx);
-        // The residual jumps when a control switches, and that jump is not a
-        // failure to make progress. Letting a globalisation veto it stalls the
-        // active set instead of resolving it.
-        if (controls_moved) {
-            for (int i = 0; i < n; ++i) {
-                x[i] += dx[i];
-            }
-        } else {
-            x = globalisation.accept(system, x, r, dx);
-        }
-    }
-    last.iterations = max_iterations + 1;
-    last.node_pressure = system.pressures(x);
-    last.well_rate = system.wellRates(x);
-    return last;
-}
-
 } // namespace Opm::NetworkSolve
 
-#endif // OPM_NETWORK_SYSTEM_HEADER_INCLUDED
+#endif // OPM_NETWORK_PRODUCTION_SYSTEM_HEADER_INCLUDED
